data/requirements.txt
deep_translator
progressbar2